Detect authentication errors in fetch_SPARQL
If SBOLQuery.login is not invoked before a query, the result comes
back as bytes instead of a dict.  This is hard to detect by the caller
and does not indicate the underlying problem. Detect the lack of
authentication in SBOLQuery.fetch_SPARQL and raise an appropriate
exception so that it is clear to the caller what has gone wrong.

Add unit tests for a query with and without login to ensure the case
is properly detected going forward.