Relax pySBOL dependency
Drop the ".postX" from the pySBOL dependency specification. Per the
post-releases section of PEP 440, there shouldn't be code changes
between post releases, so we shouldn't specify a post release in our
dependency.

Also document the block that sets up dependencies based on pip version
to explain what is being done and why it is being done.