Handle string arguments instead of lists
For some methods, it is confusing to pass a list of one
element. Passing the element alone causes a large stacktrace with
confusing errors. Modify the underlying routines to handle either
lists or strings gracefully.

Additionally, fix up query_collections() to handle a string instead of
a list when generating the variables in the query.

Add unit tests for the underlying routines and for query_collections.